refactor: simplify report service with unified filtering method
